Adding change detection training set generator and training colab notebooks
Training set generator is working but incomplete (e.g. need to be able to
highlight frames for specific inclusion as negative items).

Training is not working; at present it only attempts to load the available
data sets and is entirely untested (due to lack of working internet
connetion at time of writing).